+++ composer_local_dev/constants.py
@@ -223,6 +223,12 @@ INVALID_INT_RANGE_VALUE_ERROR = (
     "{value} is not in the range {allowed_range}."
 )
 INVALID_IMAGE_VERSION_ERROR = "Composer version must match `composer-(2.y.z|3)-airflow-a.b.c[-build.d]` pattern."
+IMAGE_VERSION_IS_NOT_SUPPORTED_ERROR = (
+    "Composer version {image_version} is not supported. Please make sure to "
+    "use existing Cloud Composer version. You can see the list of "
+    "supported versions by running:\n\n"
+    "    $ composer-dev list-available-versions"
+)
 IMAGE_TAG_DOES_NOT_EXIST_ERROR = (
     "Composer version {image_tag} seems not to be valid. Please make sure to "
     "use existing Cloud Composer version. You can see the list of "
@@ -248,6 +254,7 @@ DOCKER_NOT_AVAILABLE_ERROR = (
     "is installed and running. Error: {error}"
 )
 DOCKER_CONTAINER_MEMORY_LIMIT = "4g"
+DOCKER_CONTAINER_CPU_COUNT = 2
 NOT_MODIFIABLE_ENVIRONMENT_VARIABLES = {
     "AIRFLOW_HOME",
 }